+++ lib/src/components/material_input/material_input.dart
@@ -122,6 +122,7 @@ const String materialInputErrorKey = 'material-input-error';
       'error',
       'errorMsg',
       'label',
+      'errorRenderer',
       'floatingLabel',
       'hintText',
       'showHintOnlyOnFocus',